requirements.txt
Django>=5.1.0,<5.2.0
djangorestframework>=3.14.0,<3.15.0
Pillow>=10.0.0,<11.0.0
python-dotenv>=1.0.0,<1.1.0
